+++ back-end/.gitignore
@@ -1,5 +1,2 @@
 node_modules
-/node_modules
-
-node_modules/
 .env
