+++ back-end/package.json
@@ -4,15 +4,18 @@
   "description": "",
   "main": "gymkingheroku.ts",
   "scripts": {
-    "start": "ts-node index.ts",
+    "compile": "chmod -R 777 Models/",
+    "start": "ts-node ./src/index.ts",
     "start:api": "heroku local web",
-    "test": "npx jest"
+    "test": "npx jest --runInBand"
   },
   "keywords": [],
   "author": "Ryan Broemer - Codebusters @ University of Pretoria COS301",
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.0.1",
+    "bcryptjs": "^2.4.3",
+    "body-parser": "^1.20.0",
     "cors": "^2.8.5",
     "express": "^4.18.1",
     "pg": "^8.7.3",
