+++ back-end/src/Models/concept.glb
