+++ back-end/src/index.ts
@@ -1,3 +1,3 @@
-import { server } from "./utils/utils";
+import {server} from "./server"
 const PORT = process.env.PORT || 8081
 server.listen(PORT, () => {console.log(`Listening on ${ PORT }`);})
