+++ back-end/src/owners/owners.ts
@@ -1,9 +1,10 @@
 const express = require("express");
-const path = require("path");
 const cors = require("cors");
 const { Pool } = require("pg");
-const allowedOrigins = ["http://localhost:3000"];
-import { server } from "../employees/employees";
+const allowedOrigins = [
+  'http://localhost:3000',
+  'http://localhost:8100'
+];
 const corsOptions = {
   origin: (origin: any, callback: any) => {
     if (allowedOrigins.includes(origin) || !origin) {
@@ -39,7 +40,7 @@ function createID(length: any) {
   }
   return ID;
 }
-server
+const owners = express.Router()
   .options("*", cors(corsOptions))
   .get("/gyms/owned", cors(corsOptions), async (req: any, res: any) => {
     try {
@@ -143,4 +144,4 @@ server
       res.json(results);
     }
   });
-export { server };
+export {owners};
